añadido mockito en test Categoria Action list y Categoria service findbyid y finAll

+++ src/main/java/es/toni/struts/crud/dao/CategoriaDao.java
@@ -6,7 +6,7 @@ import es.toni.struts.crud.model.Categoria;
 
 public interface CategoriaDao {
     
-	Categoria findById(Integer id);
+	Categoria findById(Long id);
 	
     List<Categoria> findAll();
